CHROMIUM: drm: bridge/analogix: Properly log AUX CH errors
The code in analogix_dp_transfer() that was supposed to print out:
  AUX CH error happened

Was actually dead code.  That's because the previous check (whether
the interrupt status indicated any errors) would have hit for all
errors anyway.

Let's combine the two error checks so we can actually see AUX CH
errors.  We'll also downgrade the message to a warning since some of
these types of errors might be expected for some displays.  If this
gets too noisy we can downgrade again to debug.

BUG=chrome-os-partner:60130
TEST=Find a unit that has aux errors and see debug prints

Change-Id: I26fe85077fcab83af73d932c945e8ef8b0f696e2
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/433453
Tested-by: 征增 王 <wzz@rock-chips.com>
Reviewed-by: 征增 王 <wzz@rock-chips.com>