CHROMIUM: [media] rockchip-vpu: Reset raw format when coded format changes
Support for raw format and resolution depends on selected coded format,
so whenever S_FMT is called on the queue that works on coded data, we
need to make sure that raw video queue has a valid format. Current code
does not do so.

To fix the problem, let's reset raw format to default whenever coded
format changes just to stay on the safe side and keep driver state sane.
The exact format does not matter at this point, since the API mandates
the user to always set the raw format after coded format.

BUG=chrome-os-partner:62115
TEST=Run VDA unittest on Kevin

Change-Id: I2f3bcc20a814b975d50c3cb41476061ecf88e85b
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/435062
Reviewed-by: Pawel Osciak <posciak@chromium.org>