CHROMIUM: drm/img-rogue: Don't fail probe if no cpu_thermal found
The cpu_thermal node is used to find a temperature sensor for use
during devfreq cooling.  If the node is not found, this shouldn't
be fatal - it just means that the driver can't cool itself properly.
So, don't fail probe, just warn and keep going.

This is useful in early board bring-up, where the GPU driver is
ready before cpufreq and cpu cooling.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:60805
TEST=build chromeos-4.4 for rowan and boot to UI

Change-Id: I2b8cce3d69fce276aa2248dda26728734ae72da1
Reviewed-on: https://chromium-review.googlesource.com/444471
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>