CHROMIUM: arm64: dts: mt8173-oak: Force vgp3 to 1.8 V
VGP3 (aka vcamaf) is one of the variable voltage LDO regulators in
the MT6397 PMIC.

On oak, its output is used as the SDIO IO voltage for mmc3, which is the
SDIO port used for the WiFi/BT module.

This module only supports 1.8 V IO, so just force this regulator to be
a fixed 1.8 V supply rather than a variable output regulator with a
narrow 1.62~1.98 V range.

This fix is needed to be compatible with upstream commit fa93fd4ecc9c
("regulator: core: Ensure we are at least in bounds for our constraints").

The vgp3 initial voltage is 2.8 V, the above commit will force vgp3 back
down to the max voltage in its specified range; before this fix, that
voltage (1.98 V) was not valid for this regulator.  This patch fixes
this by supplying a valid max voltage of 1.8 V).

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:63054
TEST=boot elm to UI, no messages about vcamaf or vgp3.
TEST=WiFi works, can scan APs

Change-Id: I97787a32c8b2edf57b0a11eebaa8fd4ffd25b453
Reviewed-on: https://chromium-review.googlesource.com/444705
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>