UPSTREAM: drm/rockchip: cdn-dp: Fix error handling
It is likely that both 'clk_disable_unprepare()' should be called if
'pm_runtime_get_sync()' fails.

Add a new label for that, because 'err_set_rate' is not meaningful in this
case.

Add a missing call to 'pm_runtime_put()'.

Fixes: 1a0f7ed3abe2 ("drm/rockchip: cdn-dp: add cdn DP support for rk3399")

BUG=None
TEST=Check DP display, it works well

Signed-off-by: Christophe JAILLET <christophe.jaillet@wanadoo.fr>
Acked-by: Mark Yao <mark.yao@rock-chips.com>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/20170220070815.23096-1-christophe.jaillet@wanadoo.fr

(cherry picked from git://anongit.freedesktop.org/drm-misc drm-misc-next
 commit efe0220fc2d2cc1cbd6c663dd3d652ac2b9afd1a)
Change-Id: I30bae900a98fe7159728043aa9321d7d5f9573ec
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/451046
Reviewed-by: Sean Paul <seanpaul@chromium.org>