UPSTREAM: drm/rockchip: dw-mipi-dsi: fix escape clock rate
This clock rate is derived from the PHY PLL, so it should be calculated
dynamically.  This calculation is the same as that used by the vendor
kernel and ensures that the escape clock runs at <20MHz as required by
the MIPI specification.

BUG=chrome-os-partner:62134
TEST=check mipi screen, it can display normally

Signed-off-by: John Keeping <john@metanate.com>
Reviewed-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/20170224125506.21533-14-john@metanate.com

(cherry picked from git://anongit.freedesktop.org/drm-misc drm-misc-next
 commit 1bef24bae2ed4d9eabc5cde33e293b2fe95796ef)
Change-Id:I3079f9109d6fe1ab610e893b5a5c313249086a4a
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/430609
Reviewed-by: Brian Norris <briannorris@chromium.org>