UPSTREAM: drm/rockchip: dw-mipi-dsi: properly configure PHY timing
These values are specified as constant time periods but the PHY
configuration is in terms of the current lane byte clock so using
constant values guarantees that the timings will be outside the
specification with some display configurations.

Derive the necessary configuration from the byte clock in order to
ensure that the PHY configuration is correct.

BUG=chrome-os-partner:62134
TEST=check mipi screen, it can display normally

Signed-off-by: John Keeping <john@metanate.com>
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Link: http://patchwork.freedesktop.org/patch/msgid/20170224125506.21533-17-john@metanate.com

(cherry picked from git://anongit.freedesktop.org/drm-misc drm-misc-next
 commit 3fdfb4f170df4d113a3b6663ebf115e10601d0e9)
Change-Id:I0e6ddd6395a65eb9a1723639339af5ec7c9d1976
Signed-off-by: Chris Zhong <zyw@rock-chips.com>
Reviewed-on: https://chromium-review.googlesource.com/431273
Reviewed-by: Brian Norris <briannorris@chromium.org>