CHROMIUM: ARM64: dts: rk3399-gru: Assign VOP clocks with VOP IOMMUs
VOP clocks need to be set up properly for VOP IOMMUs to work, however
currently they are only initialized when VOPs themselve are probed.
With planned changes, the IOMMUs will have to be accessed before that,
so we need to make sure the clocks are set up as well.

BUG=b:36082061
BUG=b:35586595
TEST=Boot Kevin to UI.

Suggested-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Change-Id: I4e7a9fd7ed98e9986facfc4fffe8caf123c5ca4d
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/454498
Commit-Ready: Jeffy Chen <jeffy.chen@rock-chips.com>
Tested-by: Jeffy Chen <jeffy.chen@rock-chips.com>
Reviewed-by: Douglas Anderson <dianders@chromium.org>