CHROMIUM: usb: dwc3: rockchip: Fix pm runtime code
pm_runtime_get_sync(dev) is called on probe and pm_runtime_put_sync(dev)
on remove. Result is that pm runtime is always active in between, the
runtime suspend/resume functions are never called, and the usb clocks
are always active even if nothing is connected to the Type-C ports.

pm_runtime_get_sync() is supposed to be called from chip drivers if/when
needed, so drop the call and enable runtime suspend earlier to solve the
problem. Also, disable runtime suspend in the remove function only after
the call to dwc3_rockchip_of_depopulate().

BUG=b:36250586
TEST=Insert/remove Type-C cable and observe pm runtime working

Change-Id: Ic999d9c2f7c08d2405e782d4dee13fa9d3f451c9
Signed-off-by: Guenter Roeck <groeck@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/455419
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>