CHROMIUM: cros_ec: fail early if we cannot identify the EC
If we cannot communicate with the EC chip to detect the protocol version
and its features, it's very likely useless to continue.
Else we will commit all kind of uninformed mistakes (using the wrong
protocol, the wrong buffer size, mixing the EC with other chips).

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=b:35648259
TEST=run on Eve without FP MCU and see the 'cros-ec-spi' failing
early with 'Cannot identify the EC: error -110' and the fpc1020 module
starting.

Change-Id: I4b667f9e11e0da42097c7cd4107b3937203c09e1
Reviewed-on: https://chromium-review.googlesource.com/455784
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>