CHROMIUM: drm/rockchip: YUV overlays BT.601 color conversion.
Currently NV12 hardware overlays scheduled with atomic interface are
converted to RGB using a color space conversion different than BT.601.

The result is that colors of NV12 buffers composited with Mali don't
match colors of YUV hardware overlays.

Running modetest with an NV12 plane also shows colors are incorrect
if compared to an RGB overlay that represents the same color pattern.

This CL adds support for YUV2YUV color space conversion (CSC) module on
rockchip and enables the YUV2RGB part that allows to specify a 4x3
matrix to convert the colors.

The matrix is set to BT.601 coefficients to align to Mali and to what
modetest expects.

BUG=b:36752417
TEST=modetest -M rockchip -s 34@30:2400x1600 -P 30:640x480+650+10@NV12 -P 30:640x480+10+10@XR24

Change-Id: Ib3c30f457360b582e3cdfd1bd64d0652ee222fdf
Reviewed-on: https://chromium-review.googlesource.com/469392
Commit-Ready: Daniele Castagna <dcastagna@chromium.org>
Tested-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Kristian H. Kristensen <hoegsberg@chromium.org>